feat: bottom tab plcaeholder
